Refactor transactions
Split transactions into seperate file
